ai_pc_webui_full/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.3
openai==1.57.4
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
pywebview==5.3.2
